Remove post 4 but don't remove it from public
The purpose of the last 4 commits is to determine what happens if you
erroneously add the public directory to source control, push, then
remove content without removing it from the public directory.

The result is that extra file inthe public directory is not deleted
when the GitHub Pages workflow runs. This is expected. The live site
still contains content that was deleted from source control.w